Refraction w/ total internal reflection

+++ include/rt/math/vec3.hpp
@@ -338,27 +338,27 @@ constexpr float dot(const Vec3<A>& u, const Vec3<B>& v) noexcept {
  *
  * Randomization is based off a non-uniform cosine-based distribution where directions closer to the normal
  * are prioritized.
- * @param normal Outward-facing normal vector (the direction represented by the returned unit vector will never oppose this).
+ * @param normal Incident ray-facing normal vector (not to be confused with the outward-facing normal).
  * @return Randomly-generated unit vector that adheres to the above criteria.
  */
 [[nodiscard]] uvec3 scatter_uvec3(const uvec3& normal);
 
 /**
  * @brief Generates a direction unit vector of another vector, reflected (to trace ray reflection).
- * @param v Vector to be reflected.
- * @param normal Reflection will be in the direction of the outward-facing normal.
+ * @param v Direction-vector to be reflected.
+ * @param normal Incident ray-facing normal vector (not to be confused with the outward-facing normal).
  * @return Reflected direction vector that adheres to the above criteria.
  */
-[[nodiscard]] uvec3 reflect_uvec3(const vec3& v, const uvec3& normal);
+[[nodiscard]] uvec3 reflect_uvec3(const uvec3& v, const uvec3& normal);
 
 /**
- * @brief Generates a direction unit vector of another vector, refracted (to trace ray reflection).
- * @param v Vector to be reflected.
- * @param normal Refraction will be in the opposite direction of the outward-facing normal.
+ * @brief Generates a direction unit vector of another vector, refracted (to trace ray refraction) or reflected if total internal reflection applies.
+ * @param v Direction-vector to be refracted.
+ * @param normal Incident ray-facing normal vector (not to be confused with the outward-facing normal).
  * @param eta Refractive index of the incident medium (medium before ray-object intersect).
  * @param eta_prime Refractive index of the new medium (medium after ray-object intersect).
- * @return Reflected direction vector that adheres to the above criteria.
+ * @return Refracted direction vector that adheres to the above criteria (unless reflected due to total internal reflection).
  */
-[[nodiscard]] uvec3 refract_uvec3(const vec3& v, const uvec3& normal, float eta, float eta_prime);
+[[nodiscard]] uvec3 refract_uvec3(const uvec3& v, const uvec3& normal, float eta, float eta_prime);
 
 #endif
